+++ src/constant/errorCode.js
@@ -1,0 +1,11 @@
+const errorCode = {
+  OK: 200,
+  CREATED: 201,
+  BAD_REQUEST: 400,
+  UNAUTHORIZED: 401,
+  FORBIDDEN: 403,
+  NOT_FOUND: 404,
+  INTERNAL_SERVER: 500
+}
+
+module.exports = errorCode
