+++ src/constant/index.js
@@ -1,0 +1,10 @@
+const RoleShop = {
+  SHOP: 'SHOP',
+  WRITER: 'WRITER',
+  EDITOR: 'EDITOR',
+  ADMIN: 'ADMIN'
+}
+
+module.exports = {
+  RoleShop
+}
